Move bin/run file
