+++ lib/kaba.js
@@ -1,6 +1,6 @@
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
 const FriendlyErrorsWebpackPlugin = require('friendly-errors-webpack-plugin');
-const cli = require("../cli/cli");
+const cli = require("../bin/cli");
 const fs = require("fs-extra");
 const path = require("path");
 const postCssReporter = require("postcss-reporter");
